requirements.txt
pandas
seaborn
matplotlib
psycopg2
python-dotenv

